Basic connection page functionality
Styles the connection page to make it a bit better and easier to use.

Adds a "show password" checkbox to the page and defaults the input type
for the password to "password" to hide characters by default.

Adds focus indicators to input fields (hooray a11y)

Adds some instructions to the page for users of init()

Semantic HTML, wew